+++ tests/tickets/9456/1.js
@@ -1,4 +1,4 @@
-/* bender-tags: editor,unit,pastefromword */
+/* bender-tags: editor,unit,clipboard,pastefromword */
 /* bender-ckeditor-plugins: clipboard,pastefromword,format,ajax */
 /* global assertPasteEvent */